Reorganized solution Made interface for lib-users more straightforward (constructor, statistics)
