+++ src/sharp-arduino-serial-packet-lib/sharp-arduino-serial-packet-lib/Statistics.cs
@@ -1,0 +1,8 @@
+﻿namespace sharp_arduino_serial_packet_lib
+{
+    public class Statistics
+    {
+        public int ReceivedPackets { get; set; }
+        public int CorruptPackets { get; set; }
+    }
+}
