+++ src/sharp-arduino-serial-packet-lib/sharp-arduino-serial-packet-lib/sharp-arduino-serial-packet-lib.csproj
@@ -43,12 +43,14 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Defines.cs" />
+    <Compile Include="Defaults.cs" />
+    <Compile Include="ExtensionsMethods.cs" />
     <Compile Include="Packet.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="SerialPortManager.cs" />
+    <Compile Include="SerialPortReaderWriter\SerialPortManager.cs" />
     <Compile Include="SerialReaderWriter.cs" />
-    <Compile Include="SerialSettings.cs" />
+    <Compile Include="SerialPortReaderWriter\SerialSettings.cs" />
+    <Compile Include="Statistics.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="HelpStuff.txt" />
